+++ Assets/Scripts/StoneMechanics/StoneTypes.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: fe265dc9883eca74e9950a919611190c
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
