+++ Assets/Scripts/StoneMechanics/StoneTypes/BounceStone.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 3c3ef885e6af16240bda0b926babe01f
+guid: a057a4a4f4fb4344480c5c231537f647
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
