+++ Assets/Scripts/StoneMechanics/StoneTypes/ExplosionStone.cs
@@ -1,0 +1,33 @@
+using UnityEngine;
+
+namespace StoneTypes
+{
+    public class ExplosionStone : StoneBehaviour
+    {
+        public ExplosionStone() : base()
+        {
+            this._stoneSpeed = 10f;
+            this.stoneTextureLocation = "Stones/explosion-stone";
+        }
+
+        public override void ThrowStone(Rigidbody2D stoneBody)
+        {
+            base.ThrowStone(stoneBody);
+        }
+
+        public override void OnCollisionEnter(Collision collision)
+        {
+            base.OnCollisionEnter(collision);
+        }
+
+        public override void Update()
+        {
+            base.Update();
+        }
+
+        public override void FixedUpdate()
+        {
+            base.FixedUpdate();
+        }
+    }
+}
