+++ Assets/Scripts/StoneMechanics/StoneTypes/ExplosionStone.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 5ce2e568bf374c743845be2ecaaec3b1
+guid: f8b22b53a09c57f4abfa3277c18a15d2
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
