+++ Assets/Scripts/StoneMechanics/StoneTypes/FireStone.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: fb05d1abe2c91ba4c9db62acac5e1de8
+guid: f862fc07e278fb4468883d1ba1993984
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
