+++ Assets/Scripts/StoneMechanics/StoneTypes/JokerStone.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: a59be3acb59c5e447ad6cd505024b8a7
+guid: cdd3b15ac4edf1d4e96c1694cf99bbf3
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
