+++ Assets/Scripts/StoneMechanics/StoneTypes/MindControlStone.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 4ae1b6d2259cae74583894adef0470de
+guid: 9785223fcd831f740ae69cb48a9fa864
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
