+++ Assets/Scripts/StoneMechanics/StoneTypes/NormalStone.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 4e6f12d4cb565c648bc914660df44a38
+guid: 8b2224c54f70f004089680aa46deb142
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
