+++ Assets/Scripts/StoneMechanics/StoneTypes/TeleportStone.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 4860a2458d3e72540a4942806dc75a23
+guid: 9f7d86f4a55916544920d54478a34053
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
